requirements.txt
Django==1.10.6
psycopg2==2.6.2
djangorestframework==3.5.1
django-filter==1.0.1
